(api) add: Create expenses endpoint
