+++ expenses-node/index.js
@@ -2,9 +2,10 @@ require('dotenv').config()
 
 const express = require('express')
 const mongoose = require('mongoose')
+const bodyParser = require('body-parser');
 
 const test = require('./routes/test.route')
-const expenses = require('./routes/expense.route')
+const expenses = require('./routes/expenses.route')
 const app = express()
 
 const port = process.env.PORT || 4000
@@ -18,6 +19,10 @@ mongoose.Promise = global.Promise
 const db = mongoose.connection
 
 db.on('error', console.error.bind(console, 'MongoDB connection error:'))
+
+// bodyParser config
+app.use(bodyParser.json())
+app.use(bodyParser.urlencoded({ extended: false }))
 
 // Routes
 app.use('/', test)
