--- expenses-node/routes/expense.route.js
@@ -1,7 +1,0 @@
-const express = require('express')
-const router = express.Router()
-const expenseController = require('../controllers/expense.controller')
-
-router.get('/', expenseController.index);
-
-module.exports = router
