+++ expenses-node/routes/expenses.route.js
@@ -1,0 +1,8 @@
+const express = require('express')
+const router = express.Router()
+const expensesController = require('../controllers/expenses.controller')
+
+router.get('/', expensesController.index);
+router.post('/', expensesController.create);
+
+module.exports = router
